Move dynamic plugins to a subpackage
